+++ ui/packages/shared/components/src/KeyDownContext/index.tsx
@@ -11,10 +11,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {useEffect, useState} from 'react';
+import {createContext, ReactNode, useContext, useEffect, useMemo, useState} from 'react';
 
-const useIsShiftDown = (): boolean => {
-  const [isShiftDown, setIsShiftDown] = useState(false);
+export interface KeyDownState {
+  isShiftDown: boolean;
+}
+
+const DEFAULT_VALUE = {
+  isShiftDown: false,
+};
+
+const KeyDownContext = createContext<KeyDownState>(DEFAULT_VALUE);
+
+export const KeyDownProvider = ({
+  children,
+}: {
+  children: ReactNode;
+  value?: KeyDownState;
+}): JSX.Element => {
+  const [isShiftDown, setIsShiftDown] = useState<boolean>(false);
 
   useEffect(() => {
     const handleShiftDown = (event: {keyCode: number}): void => {
@@ -39,7 +54,17 @@ const useIsShiftDown = (): boolean => {
     };
   }, []);
 
-  return isShiftDown;
+  const value = useMemo(() => ({isShiftDown}), [isShiftDown]);
+
+  return <KeyDownContext.Provider value={value}>{children}</KeyDownContext.Provider>;
 };
 
-export default useIsShiftDown;
+export const useKeyDown = (): KeyDownState => {
+  const context = useContext(KeyDownContext);
+  if (context == null) {
+    return DEFAULT_VALUE;
+  }
+  return context;
+};
+
+export default KeyDownContext;