+++ ui/packages/shared/components/src/TextWithTooltip/index.tsx
@@ -11,4 +11,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './useURLState';
+import {Tooltip} from 'react-tooltip';
+import 'react-tooltip/dist/react-tooltip.css';
+
+import {cutToMaxStringLength} from '@parca/functions/string';
+
+interface Props {
+  text: string;
+  maxTextLength: number;
+  id: string;
+}
+
+const TextWithTooltip = ({text, maxTextLength, id}: Props): JSX.Element => {
+  return (
+    <>
+      <span id={id} data-tooltip-content={text}>
+        {cutToMaxStringLength(text, maxTextLength)}
+      </span>
+      <Tooltip anchorId={id} />
+    </>
+  );
+};
+
+export default TextWithTooltip;